add program requirement table and selection
