+++ api/controllers/widget.js
@@ -21,7 +21,7 @@ const student_name = 'PreCustomer';
 const WidgetProcessTranscriptV2 = asyncHandler(async (req, res, next) => {
   const {
     params: { category, language },
-    body: { courses, table_data_string_taiger_guided }
+    body: { courses, requirementIds }
   } = req;
 
   const { Credentials } = await getTemporaryCredentials(
@@ -29,9 +29,6 @@ const WidgetProcessTranscriptV2 = asyncHandler(async (req, res, next) => {
   );
 
   const stringified_courses = JSON.stringify(JSON.stringify(courses));
-  const stringified_courses_taiger_guided = JSON.stringify(
-    JSON.stringify(table_data_string_taiger_guided)
-  );
   const studentId = req.user._id.toString();
   try {
     // TODO: replacing requirement_ids with data from frontend.
@@ -42,11 +39,8 @@ const WidgetProcessTranscriptV2 = asyncHandler(async (req, res, next) => {
       student_id: studentId,
       student_name,
       language,
-      courses_taiger_guided: stringified_courses_taiger_guided,
-      requirement_ids: JSON.stringify([
-        '6716e505b5c65b932fda02a4',
-        '671ecb55015d01f7771f0fac'
-      ])
+      courses_taiger_guided: '[]',
+      requirement_ids: JSON.stringify(requirementIds)
     });
 
     const metadata = {
