+++ client/src/Demo/MyCourses/CourseWidgetBody.js
@@ -1,0 +1,685 @@
+import React, { useState } from 'react';
+import {
+  Box,
+  Button,
+  CircularProgress,
+  Card,
+  FormControl,
+  InputLabel,
+  Link,
+  MenuItem,
+  Select,
+  Typography,
+  Dialog,
+  DialogTitle,
+  DialogContentText,
+  DialogContent,
+  DialogActions,
+  Badge,
+  Tabs,
+  Tab,
+  TextField
+} from '@mui/material';
+import { DataSheetGrid, textColumn, keyColumn } from 'react-datasheet-grid';
+import { Navigate, Link as LinkDom, useParams } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+import 'react-datasheet-grid/dist/style.css';
+
+import { study_group } from '../Utils/contants';
+import ErrorPage from '../Utils/ErrorPage';
+import ModalMain from '../Utils/ModalHandler/ModalMain';
+import { is_TaiGer_role } from '../Utils/checking-functions';
+import {
+  WidgetanalyzedFileDownload,
+  WidgetTranscriptanalyser,
+  WidgetTranscriptanalyserV2
+} from '../../api';
+import { TabTitle } from '../Utils/TabTitle';
+import DEMO from '../../store/constant';
+import { useAuth } from '../../components/AuthProvider';
+import Loading from '../../components/Loading/Loading';
+import { a11yProps, CustomTabPanel } from '../../components/Tabs';
+import { DataGrid, GridToolbar } from '@mui/x-data-grid';
+import CourseAnalysisConfirmDialog from './CourseAnalysisConfirmDialog';
+
+export default function CourseWidgetBody({ programRequirements }) {
+  const { user } = useAuth();
+  const { student_id } = useParams();
+  const { t } = useTranslation();
+
+  const [filters, setFilters] = useState({});
+  const [rowSelectionModel, setRowSelectionModel] = React.useState([]);
+
+  let [statedata, setStatedata] = useState({
+    error: '',
+    isLoaded: true,
+    coursesdata: [
+      {
+        course_chinese: '電子學',
+        course_english: 'Electronics',
+        credits: '3',
+        grades: 'B'
+      }
+    ],
+    analysis: {},
+    analysisSuccessModalWindowOpen: false,
+    success: false,
+    student: null,
+    file: '',
+    study_group: '',
+    analysis_language: '',
+    analyzed_course: '',
+    expand: true,
+    isAnalysing: false,
+    isUpdating: false,
+    isDownloading: false,
+    res_status: 0,
+    res_modal_status: 0,
+    res_modal_message: ''
+  });
+
+  const [value, setValue] = useState(0);
+
+  const handleChangeValue = (event, newValue) => {
+    setValue(newValue);
+  };
+
+  const onChange = (new_data) => {
+    setStatedata((state) => ({
+      ...state,
+      coursesdata: new_data
+    }));
+  };
+
+  const handleChange_study_group = (e) => {
+    e.preventDefault();
+    const { value } = e.target;
+    setStatedata((state) => ({
+      ...state,
+      study_group: value
+    }));
+  };
+
+  const handleChange_analysis_language = (e) => {
+    e.preventDefault();
+    const { value } = e.target;
+    setStatedata((state) => ({
+      ...state,
+      analysis_language: value
+    }));
+  };
+
+  const ConfirmError = () => {
+    setStatedata((state) => ({
+      ...state,
+      res_modal_status: 0,
+      res_modal_message: ''
+    }));
+  };
+
+  const onAnalyse = () => {
+    if (statedata.study_group === '') {
+      alert('Please select study group');
+      return;
+    }
+    setStatedata((state) => ({
+      ...state,
+      isAnalysing: true
+    }));
+    WidgetTranscriptanalyser(
+      statedata.study_group,
+      statedata.analysis_language,
+      statedata.coursesdata,
+      []
+    ).then(
+      (resp) => {
+        const { data, success } = resp.data;
+        const { status } = resp;
+        if (success) {
+          setStatedata((state) => ({
+            ...state,
+            isLoaded: true,
+            analysis: data,
+            analysisSuccessModalWindowOpen: true,
+            success: success,
+            isAnalysing: false,
+            res_modal_status: status
+          }));
+        } else {
+          setStatedata((state) => ({
+            ...state,
+            isLoaded: true,
+            isAnalysing: false,
+            res_modal_status: status,
+            res_modal_message:
+              'Make sure that you updated your courses and select the right target group and language!'
+          }));
+        }
+      },
+      (error) => {
+        setStatedata((state) => ({
+          ...state,
+          isLoaded: true,
+          isAnalysing: false,
+          error,
+          res_modal_status: 500,
+          res_modal_message:
+            'Make sure that you updated your courses and select the right target group and language!'
+        }));
+      }
+    );
+  };
+
+  const setModalHide = () => {
+    setStatedata((state) => ({
+      ...state,
+      modalShowAssignWindow: false
+    }));
+  };
+
+  const setModalShow2 = () => {
+    setStatedata((state) => ({
+      ...state,
+      modalShowAssignWindow: true
+    }));
+  };
+
+  const onAnalyseV2 = async () => {
+    setStatedata((state) => ({
+      ...state,
+      isAnalysing: true
+    }));
+
+    try {
+      const resp = await WidgetTranscriptanalyserV2(
+        statedata.study_group,
+        statedata.analysis_language,
+        statedata.coursesdata,
+        rowSelectionModel
+      );
+
+      const { data, success } = resp.data;
+      const { status } = resp;
+      if (success) {
+        setStatedata((state) => ({
+          ...state,
+          isLoaded: true,
+          analysis: data,
+          analysisSuccessModalWindowOpen: true,
+          success: success,
+          isAnalysing: false,
+          res_modal_status: status
+        }));
+      } else {
+        setStatedata((state) => ({
+          ...state,
+          isLoaded: true,
+          isAnalysing: false,
+          res_modal_status: status,
+          res_modal_message:
+            'Make sure that you updated your courses and select the right target group and language!'
+        }));
+      }
+    } catch (error) {
+      setStatedata((state) => ({
+        ...state,
+        isLoaded: true,
+        isAnalysing: false,
+        error,
+        res_modal_status: 500,
+        res_modal_message:
+          'Make sure that you updated your courses and select the right target group and language!'
+      }));
+    }
+  };
+
+  const onDownload = () => {
+    setStatedata((state) => ({
+      ...state,
+      isDownloading: true
+    }));
+    WidgetanalyzedFileDownload(user._id.toString()).then(
+      (resp) => {
+        // TODO: timeout? success?
+        const { status } = resp;
+        if (status < 300) {
+          const actualFileName = decodeURIComponent(
+            resp.headers['content-disposition'].split('"')[1]
+          ); //  檔名中文亂碼 solved
+          const { data: blob } = resp;
+          if (blob.size === 0) return;
+
+          var filetype = actualFileName.split('.'); //split file name
+          filetype = filetype.pop(); //get the file type
+
+          if (filetype === 'pdf') {
+            const url = window.URL.createObjectURL(
+              new Blob([blob], { type: 'application/pdf' })
+            );
+
+            //Open the URL on new Window
+            window.open(url); //TODO: having a reasonable file name, pdf viewer
+          } else {
+            //if not pdf, download instead.
+
+            const url = window.URL.createObjectURL(new Blob([blob]));
+
+            const link = document.createElement('a');
+            link.href = url;
+            link.setAttribute('download', actualFileName);
+            // Append to html link element page
+            document.body.appendChild(link);
+            // Start download
+            link.click();
+            // Clean up and remove the link
+            link.parentNode.removeChild(link);
+            setStatedata((state) => ({
+              ...state,
+              isDownloading: false
+            }));
+          }
+        } else {
+          const { statusText } = resp;
+          setStatedata((state) => ({
+            ...state,
+            isLoaded: true,
+            res_modal_status: status,
+            res_modal_message: statusText,
+            isDownloading: false
+          }));
+        }
+      },
+      (error) => {
+        setStatedata((state) => ({
+          ...state,
+          isLoaded: true,
+          error,
+          res_modal_status: 500,
+          res_modal_message: '',
+          isDownloading: false
+        }));
+      }
+    );
+  };
+
+  const handleFilterChange = (event, column) => {
+    event.preventDefault();
+    const { value } = event.target;
+    const searchParams = new URLSearchParams(window.location.search);
+    searchParams.set(column.field, value.toLowerCase());
+    setFilters((prevFilters) => ({
+      ...prevFilters,
+      [column.field]: value.toLowerCase()
+    }));
+  };
+
+  const transformedData = programRequirements.map((row) => {
+    return {
+      ...row, // Spread the original row object
+      program_name: `${row.programId[0].school} ${row.programId[0].program_name} ${row.programId[0].degree}`,
+      lang: `${row.programId[0].lang}`,
+      degree: `${row.programId[0].degree}`,
+      country: `${row.programId[0].country}`,
+      id: row._id // Map MongoDB _id to id property
+      // other properties...
+    };
+  });
+
+  const filteredRows = transformedData.filter((row) => {
+    return Object.keys(filters).every((field) => {
+      const filterValue = filters[field];
+      return (
+        filterValue === '' ||
+        row[field]?.toString().toLowerCase().includes(filterValue)
+      );
+    });
+  });
+
+  const programListColumn = [
+    {
+      field: 'program_name',
+      headerName: t('Program Name', { ns: 'common' }),
+      align: 'left',
+      headerAlign: 'left',
+      width: 450,
+      renderCell: (params) => {
+        const linkUrl = `${DEMO.PROGRAM_ANALYSIS}`;
+        return (
+          <Link
+            underline="hover"
+            to={linkUrl}
+            component={LinkDom}
+            target="_blank"
+          >
+            {params.value}
+          </Link>
+        );
+      }
+    },
+    {
+      field: 'country',
+      headerName: t('Country', { ns: 'common' }),
+      width: 80
+    },
+    { field: 'degree', headerName: t('Degree', { ns: 'common' }), width: 90 },
+    {
+      field: 'lang',
+      headerName: t('Language', { ns: 'common' }),
+      width: 90
+    },
+    {
+      field: 'updatedAt',
+      headerName: t('Last update', { ns: 'common' }),
+      width: 150
+    }
+  ];
+
+  const stopPropagation = (event) => {
+    event.preventDefault();
+    event.stopPropagation();
+  };
+
+  const closeanalysisSuccessModal = () => {
+    setStatedata((state) => ({
+      ...state,
+      analysisSuccessModalWindowOpen: false
+    }));
+  };
+  const columns = [
+    {
+      ...keyColumn('course_chinese', textColumn),
+      title: 'Courses Name Chinese'
+    },
+    {
+      ...keyColumn('course_english', textColumn),
+      title: 'Courses Name English'
+    },
+    {
+      ...keyColumn('credits', textColumn),
+      title: 'Credits'
+    },
+    { ...keyColumn('grades', textColumn), title: 'Grades' }
+  ];
+
+  if (!statedata.isLoaded) {
+    return <Loading />;
+  }
+  if (!student_id) {
+    if (!is_TaiGer_role(user)) {
+      return <Navigate to={`${DEMO.DASHBOARD_LINK}`} />;
+    }
+  }
+
+  if (statedata.res_status >= 400) {
+    return <ErrorPage res_status={statedata.res_status} />;
+  }
+  TabTitle(`Course Analyser`);
+
+  return (
+    <Box>
+      {statedata.res_modal_status >= 400 && (
+        <ModalMain
+          ConfirmError={ConfirmError}
+          res_modal_status={statedata.res_modal_status}
+          res_modal_message={statedata.res_modal_message}
+        />
+      )}
+
+      <Box
+        display="flex"
+        justifyContent="space-between"
+        alignItems="center"
+        sx={{ my: 1 }}
+      >
+        <Typography variant="h6">
+          {t('Course Analyser', { ns: 'common' })}
+        </Typography>
+        <Box>
+          <Button
+            variant="outlined"
+            component={LinkDom}
+            to={`${DEMO.KEYWORDS_EDIT}`}
+            color="primary"
+            target="_blank"
+            sx={{ mr: 2 }}
+          >
+            {t('Edit Keywords', { ns: 'common' })}
+          </Button>
+          <Button
+            variant="contained"
+            component={LinkDom}
+            to={`${DEMO.CREATE_NEW_PROGRAM_ANALYSIS}`}
+            target="_blank"
+            color="primary"
+          >
+            {t('Create New Analysis', { ns: 'common' })}
+          </Button>
+        </Box>
+      </Box>
+      <Box>
+        <Card sx={{ p: 2 }}>
+          <Typography sx={{ px: 2, pt: 2 }}>
+            1. Please fill the courses.
+          </Typography>
+          <Typography sx={{ px: 2 }}>2. Select study group</Typography>
+          <Typography sx={{ px: 2 }}>
+            3. Select language. <b>Chinese</b> is more accurate.
+          </Typography>
+          <br />
+          <DataSheetGrid
+            height={6000}
+            style={{ minWidth: '450px' }}
+            disableContextMenu={true}
+            disableExpandSelection={false}
+            headerRowHeight={30}
+            rowHeight={25}
+            value={statedata.coursesdata}
+            autoAddRow={true}
+            onChange={onChange}
+            columns={columns}
+          />
+          <br />
+          <br />
+          <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
+            <Tabs
+              value={value}
+              onChange={handleChangeValue}
+              variant="scrollable"
+              scrollButtons="auto"
+              aria-label="basic tabs example"
+            >
+              <Tab label="Calender" {...a11yProps(0)} />
+              <Tab
+                label={
+                  <Badge badgeContent={'New'} color="error">
+                    V2
+                  </Badge>
+                }
+                {...a11yProps(1)}
+              />
+            </Tabs>
+          </Box>
+          <CustomTabPanel value={value} index={0}>
+            <FormControl fullWidth>
+              <InputLabel id="select-target-group">
+                {t('Select Target Group', { ns: 'courses' })}
+              </InputLabel>
+              <Select
+                labelId="study_group"
+                label="Select target group"
+                name="study_group"
+                id="study_group"
+                onChange={(e) => handleChange_study_group(e)}
+              >
+                <MenuItem value={''}>Select Study Group</MenuItem>
+                {study_group.map((cat, i) => (
+                  <MenuItem value={cat.key} key={i}>
+                    {cat.value}
+                  </MenuItem>
+                ))}
+              </Select>
+            </FormControl>
+            <br />
+            <br />
+            <FormControl fullWidth>
+              <InputLabel id="select-language">
+                {t('Select language', { ns: 'courses' })}
+              </InputLabel>
+              <Select
+                labelId="analysis_language"
+                label={t('Select language', { ns: 'courses' })}
+                name="analysis_language"
+                id="analysis_language"
+                onChange={(e) => handleChange_analysis_language(e)}
+              >
+                <MenuItem value={''}>{t('Select language')}</MenuItem>
+                <MenuItem value={'zh'}>中文</MenuItem>
+                <MenuItem value={'en'}>English (Beta Version)</MenuItem>
+              </Select>
+            </FormControl>
+            <br />
+            <br />
+            <Button
+              color="primary"
+              variant="contained"
+              onClick={onAnalyse}
+              disabled={
+                statedata.isAnalysing ||
+                statedata.study_group === '' ||
+                statedata.analysis_language === ''
+              }
+              endIcon={
+                statedata.isAnalysing ? <CircularProgress size={24} /> : <></>
+              }
+            >
+              {statedata.isAnalysing
+                ? t('Analysing', { ns: 'courses' })
+                : t('Analyse', { ns: 'courses' })}
+            </Button>
+          </CustomTabPanel>
+          <CustomTabPanel value={value} index={1}>
+            <Box
+              display="flex"
+              justifyContent="space-between"
+              alignItems="center"
+              sx={{ my: 1 }}
+            >
+              <Typography variant="h6"></Typography>
+              <Box>
+                <Button
+                  color="primary"
+                  variant="contained"
+                  onClick={setModalShow2}
+                  disabled={!rowSelectionModel.length > 0}
+                  endIcon={
+                    statedata.isAnalysing ? (
+                      <CircularProgress size={24} />
+                    ) : (
+                      <></>
+                    )
+                  }
+                >
+                  {statedata.isAnalysing
+                    ? t('Analysing', { ns: 'courses' })
+                    : t('Analyse V2', { ns: 'courses' })}
+                </Button>
+              </Box>
+            </Box>
+            <div style={{ height: '50%', width: '100%' }}>
+              <DataGrid
+                columnHeaderHeight={130}
+                density="compact"
+                rows={filteredRows}
+                disableColumnFilter
+                disableColumnMenu
+                disableDensitySelector
+                columns={programListColumn.map((column) => ({
+                  ...column,
+                  renderHeader: () => (
+                    <Box>
+                      <Typography
+                        sx={{ my: 1 }}
+                      >{`${column.headerName}`}</Typography>
+                      <TextField
+                        size="small"
+                        type="text"
+                        placeholder={`${column.headerName}`}
+                        onClick={stopPropagation}
+                        value={filters[column.field] || ''}
+                        onChange={(event) => handleFilterChange(event, column)}
+                        sx={{ mb: 1 }}
+                      />
+                    </Box>
+                  )
+                }))}
+                initialState={{
+                  pagination: {
+                    paginationModel: { page: 0, pageSize: 20 }
+                  }
+                }}
+                keepNonExistentRowsSelected
+                onRowSelectionModelChange={(newRowSelectionModel) => {
+                  setRowSelectionModel(newRowSelectionModel);
+                }}
+                rowSelectionModel={rowSelectionModel}
+                pageSizeOptions={[10, 20, 50, 100]}
+                checkboxSelection
+                slots={{ toolbar: GridToolbar }}
+                slotProps={{
+                  toolbar: {
+                    showQuickFilter: true
+                  }
+                }}
+              />
+            </div>
+          </CustomTabPanel>
+
+          <Typography>
+            {statedata.analysis && statedata.analysis.isAnalysed ? (
+              <>
+                <Button onClick={onDownload} disabled={statedata.isDownloading}>
+                  {t('Download', { ns: 'common' })}
+                </Button>
+                <Link
+                  to={`${DEMO.INTERNAL_WIDGET_LINK(user._id.toString())}`}
+                  target="_blank"
+                  component={LinkDom}
+                >
+                  {t('View Online', { ns: 'courses' })}
+                </Link>
+              </>
+            ) : (
+              t('No analysis yet', { ns: 'courses' })
+            )}
+          </Typography>
+        </Card>
+      </Box>
+      <CourseAnalysisConfirmDialog
+        show={statedata.modalShowAssignWindow}
+        setModalHide={setModalHide}
+        onAnalyseV2={onAnalyseV2}
+      />
+      <Dialog
+        open={statedata.analysisSuccessModalWindowOpen}
+        onClose={closeanalysisSuccessModal}
+        aria-labelledby="contained-modal-title-vcenter"
+      >
+        <DialogTitle>{t('Success', { ns: 'common' })}</DialogTitle>
+        <DialogContent>
+          <DialogContentText>
+            {t('Transcript analysed successfully!')}
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button
+            color="primary"
+            variant="contained"
+            onClick={closeanalysisSuccessModal}
+          >
+            {t('Close', { ns: 'common' })}
+          </Button>
+        </DialogActions>
+      </Dialog>
+    </Box>
+  );
+}
